for_client/requirements.txt
langchain
python-dotenv
requests
beautifulsoup4
faiss-cpu
unstructured
langchain-groq
pinecone-client
sentence-transformers
langchain-community
langchain-huggingface
streamlit
fastapi 
uvicorn
openai
langchain_openai
llama-index
llama-index-core
llama-index-llms-openai
llama-index-llms-langchain